+++ js/main.js
@@ -1,5 +1,5 @@
-import {getPhotoArray} from './post-maker.js';
-import {renderThumbs, initThumbsListener} from './thumbnail-maker.js';
+import {getPhotoArray} from './render-posts.js';
+import {renderThumbs, initThumbsListener} from './render-thumbs.js';
 
 const postsData = getPhotoArray();
 renderThumbs(postsData);
